Loading meta-features and meta-labels of the avalaible sets

+++ autogoal/datasets/cifar10.py
@@ -42,7 +42,7 @@ def load(training_batches=5):
 
     X_train = np.vstack(X_train)
     X_train = np.reshape(X_train, (-1, 3, 32, 32)).transpose(0, 2, 3, 1)
-    X_train = X_train.astype('float32')
+    X_train = X_train.astype('float64')
     y_train = np.array(y_train)
 
     test_batch = datapath("cifar10") / "test_batch"
@@ -51,7 +51,7 @@ def load(training_batches=5):
         data = pickle.load(fp, encoding="bytes")
         X_test, y_test = data[b"data"], data[b"labels"]
         X_test = np.reshape(X_test, (-1, 3, 32, 32)).transpose(0, 2, 3, 1)
-        X_test = X_test.astype('float32')
+        X_test = X_test.astype('float64')
         y_test = np.array(y_test)
 
     return X_train, y_train, X_test, y_test
